pyproject.toml
[project]
name = "dtek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest-playwright>=0.7.1",
    "python-dotenv>=1.2.1",
    "tabulate>=0.9.0",
]
